dev: make Windows spawn work, for realz
This is all kind of black magic to me. But /s must come before /c, otherwise nothing works. And even if you do that, it falls down on cmd's that have spaces in their name (even when wrapped in quotes), unless you remove the /s. Also, why join the args, that's just unnecessary and I think breaks the verbatim argument handling thing.

So this works at least. Not entirely sure why, but it does.