Fix some weird profile behavior
CouchLogin.anonymous() doesn't quite work.  And it's broken in a subtle
and weird way that makes it take over the last-used session.  Something is
very fucked in there.

For now, use the adminCouch for queries that don't need a token.  This has
the added benefit of making the admin login session stay a bit fresher.