Removed autofocus from search box
When browsing packages, a keyboard focused user is much more likely to
scroll down with the keyboard with the down arrow than start typing into
the search box. Autofocusing the search box thwarts the user's attemts
at this and they have to remove the focus from the search box first.

Closes #289